requirements.txt
Flask==2.3.3
PyYAML==6.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.0.1
